@@ -20,7 +20,8 @@
     "issues": "https://plugins.studer-raimann.ch/goto.php?target=uihk_srsu_PLNOTIFICATION"
   },
   "require": {
-    "php": ">=5.6",
+    "php": ">=7.0",
+    "srag/activerecordconfig": ">=0.1.0",
     "srag/custominputguis": ">=0.1.0",
     "srag/dic": ">=0.1.0",
     "srag/librariesnamespacechanger": ">=0.1.0",
