@@ -2,8 +2,11 @@
 
 namespace srag\Plugins\Notifications4Plugins\Notification;
 
+use ilDateTime;
+use ilNotifications4PluginsConfigGUI;
 use ilNotifications4PluginsPlugin;
 use srag\DIC\Notifications4Plugins\DICTrait;
+use srag\Plugins\Notifications4Plugins\Notification\Language\NotificationLanguage;
 use srag\Plugins\Notifications4Plugins\Utils\Notifications4PluginsTrait;
 
 /**
@@ -27,7 +30,7 @@ final class Repository {
 	/**
 	 * @return self
 	 */
-	public static function getInstance()/*: self*/ {
+	public static function getInstance(): self {
 		if (self::$instance === null) {
 			self::$instance = new self();
 		}
@@ -50,16 +53,41 @@ final class Repository {
 	public function deleteNotification(Notification $notification)/*: void*/ {
 		$notification->delete();
 
-		foreach ($notification->getNotificationLanguages() as $language) {
-			$language->delete();
+		foreach (self::notificationLanguage()->getLanguagesForNotification($notification->getId()) as $language) {
+			self::notificationLanguage()->deleteLanguage($language);
 		}
+	}
+
+
+	/**
+	 * @param Notification $notification
+	 *
+	 * @return Notification
+	 */
+	public function duplicateNotification(Notification $notification): Notification {
+		/**
+		 * @var Notification $cloned_notification
+		 */
+
+		$cloned_notification = $notification->copy();
+
+		$cloned_notification->setTitle($cloned_notification->getTitle() . " (" . self::plugin()
+				->translate("duplicated", ilNotifications4PluginsConfigGUI::LANG_MODULE_NOTIFICATIONS4PLUGIN) . ")");
+
+		$languages = [];
+		foreach (self::notificationLanguage()->getLanguagesForNotification($notification->getId()) as $language) {
+			$languages[$language->getLanguage()] = self::notificationLanguage()->duplicateLanguage($language);
+		}
+		$cloned_notification->setLanguages($languages);
+
+		return $cloned_notification;
 	}
 
 
 	/**
 	 * @return Factory
 	 */
-	public function factory()/*: Factory*/ {
+	public function factory(): Factory {
 		return Factory::getInstance();
 	}
 
@@ -67,10 +95,10 @@ final class Repository {
 	/**
 	 * @return array
 	 */
-	public function getArrayForSelection()/*: array*/ {
+	public function getArrayForSelection(): array {
 		$notifications = $this->getNotifications();
 
-		$array = array();
+		$array = [];
 
 		foreach ($notifications as $notification) {
 			$array[$notification->getName()] = $notification->getTitle() . " (" . $notification->getName() . ")";
@@ -83,19 +111,21 @@ final class Repository {
 	/**
 	 * @return array
 	 */
-	public function getArrayForTable()/*: array*/ {
-		$data = array();
+	public function getArrayForTable(): array {
+		$data = [];
 
 		$notifications = $this->getNotifications();
 
 		foreach ($notifications as $notification) {
-			$row = array();
+			$row = [];
 			$row["id"] = $notification->getId();
 			$row["title"] = $notification->getTitle();
 			$row["name"] = $notification->getName();
 			$row["description"] = $notification->getDescription();
 			$row["default_language"] = $notification->getDefaultLanguage();
-			$row["languages"] = implode(", ", $notification->getLanguages());
+			$row["languages"] = implode(", ", array_map(function (NotificationLanguage $language): string {
+				return $language->getLanguage();
+			}, $notification->getLanguages()));
 			$data[] = $row;
 		}
 
@@ -108,13 +138,12 @@ final class Repository {
 	 *
 	 * @return Notification|null
 	 */
-	public function getNotificationById(/*int*/
-		$id)/*: ?Notification*/ {
+	public function getNotificationById(int $id)/*: ?Notification*/ {
 		/**
 		 * @var Notification|null $notification
 		 */
 
-		$notification = Notification::where(array( "id" => $id ))->first();
+		$notification = Notification::where([ "id" => $id ])->first();
 
 		return $notification;
 	}
@@ -125,13 +154,12 @@ final class Repository {
 	 *
 	 * @return Notification|null
 	 */
-	public function getNotificationByName(/*string*/
-		$name)/*: ?Notification*/ {
+	public function getNotificationByName(string $name)/*: ?Notification*/ {
 		/**
 		 * @var Notification|null $notification
 		 */
 
-		$notification = Notification::where(array( "name" => $name ))->first();
+		$notification = Notification::where([ "name" => $name ])->first();
 
 		return $notification;
 	}
@@ -140,7 +168,7 @@ final class Repository {
 	/**
 	 * @return Notification[]
 	 */
-	public function getNotifications()/*: array*/ {
+	public function getNotifications(): array {
 		/**
 		 * @var Notification[] $notifications
 		 */
@@ -155,7 +183,7 @@ final class Repository {
 	 * @param Notification $notification
 	 */
 	public function storeInstance(Notification $notification)/*: void*/ {
-		$date = date("Y-m-d H:i:s");
+		$date = new ilDateTime(time(), IL_CAL_UNIX);
 
 		if (empty($notification->getId())) {
 			$notification->setCreatedAt($date);
@@ -165,8 +193,10 @@ final class Repository {
 
 		$notification->store();
 
-		foreach ($notification->getNotificationLanguages() as $language) {
-			$language->store();
+		foreach ($notification->getLanguages() as $language) {
+			$language->setNotificationId($notification->getId());
+
+			self::notificationLanguage()->storeInstance($language);
 		}
 	}
 
@@ -174,7 +204,7 @@ final class Repository {
 	/**
 	 * @return UI
 	 */
-	public function ui()/*: UI*/ {
+	public function ui(): UI {
 		return UI::getInstance();
 	}
 }
