@@ -3,22 +3,23 @@
 require_once __DIR__ . "/../vendor/autoload.php";
 
 use srag\DIC\Notifications4Plugins\Util\LibraryLanguageInstaller;
+use srag\Plugins\Notifications4Plugins\Notification\Language\NotificationLanguage;
 use srag\Plugins\Notifications4Plugins\Notification\Notification;
-use srag\Plugins\Notifications4Plugins\Notification\NotificationLanguage;
 use srag\Plugins\Notifications4Plugins\Utils\Notifications4PluginsTrait;
 use srag\RemovePluginDataConfirm\Notifications4Plugins\PluginUninstallTrait;
 
 /**
  * Class ilNotifications4PluginsPlugin
  *
+ * @author studer + raimann ag - Team Custom 1 <support-custom1@studer-raimann.ch>
  * @author Stefan Wanzenried <sw@studer-raimann.ch>
  */
 class ilNotifications4PluginsPlugin extends ilUserInterfaceHookPlugin {
 
 	use PluginUninstallTrait;
 	use Notifications4PluginsTrait;
-	const PLUGIN_ID = 'notifications4pl';
-	const PLUGIN_NAME = 'Notifications4Plugins';
+	const PLUGIN_ID = "notifications4pl";
+	const PLUGIN_NAME = "Notifications4Plugins";
 	const PLUGIN_CLASS_NAME = self::class;
 	const REMOVE_PLUGIN_DATA_CONFIRM_CLASS_NAME = Notifications4PluginsConfirm::class;
 	/**
@@ -30,7 +31,7 @@ class ilNotifications4PluginsPlugin extends ilUserInterfaceHookPlugin {
 	/**
 	 * @return self
 	 */
-	public static function getInstance() {
+	public static function getInstance(): self {
 		if (self::$instance === null) {
 			self::$instance = new self();
 		}
@@ -50,7 +51,7 @@ class ilNotifications4PluginsPlugin extends ilUserInterfaceHookPlugin {
 	/**
 	 * @return string
 	 */
-	public function getPluginName() {
+	public function getPluginName(): string {
 		return self::PLUGIN_NAME;
 	}
 
