@@ -3,7 +3,7 @@
 require_once __DIR__ . "/vendor/srag/dic/src/PHPVersionChecker.php";
 
 $id = "notifications4pl";
-$version = "3.1.0";
+$version = "3.1.1";
 $ilias_min_version = "5.3.14";
 $ilias_max_version = "5.4.999";
 $responsible = "studer + raimann ag - Team Custom 1";
