@@ -2,6 +2,7 @@
 
 namespace srag\Plugins\Notifications4Plugins\Notification;
 
+use ilNotifications4PluginsConfigGUI;
 use ilNotifications4PluginsPlugin;
 use ilSelectInputGUI;
 use srag\CustomInputGUIs\Notifications4Plugins\PropertyFormGUI\PropertyFormGUI;
@@ -29,7 +30,7 @@ final class UI {
 	/**
 	 * @return self
 	 */
-	public static function getInstance()/*: self*/ {
+	public static function getInstance(): self {
 		if (self::$instance === null) {
 			self::$instance = new self();
 		}
@@ -52,16 +53,17 @@ final class UI {
 	 *
 	 * @return array
 	 */
-	public function templateSelection(/*string*/
-		$post_key, array $placeholder_types)/*: array*/ {
+	public function templateSelection(string $post_key, array $placeholder_types): array {
 		return [
 			$post_key => [
 				PropertyFormGUI::PROPERTY_CLASS => ilSelectInputGUI::class,
 				PropertyFormGUI::PROPERTY_REQUIRED => true,
 				PropertyFormGUI::PROPERTY_OPTIONS => [ "" => "" ] + self::notification()->getArrayForSelection(),
-				"setTitle" => self::plugin()->translate("template_selection", "", [ ilNotifications4PluginsPlugin::PLUGIN_NAME ]),
+				"setTitle" => self::plugin()
+					->translate("template_selection", ilNotifications4PluginsConfigGUI::LANG_MODULE_NOTIFICATIONS4PLUGIN, [ ilNotifications4PluginsPlugin::PLUGIN_NAME ]),
 				"setInfo" => self::output()->getHTML([
-					self::plugin()->translate("template_selection_info", "", [ ilNotifications4PluginsPlugin::PLUGIN_NAME ]),
+					self::plugin()
+						->translate("template_selection_info", ilNotifications4PluginsConfigGUI::LANG_MODULE_NOTIFICATIONS4PLUGIN, [ ilNotifications4PluginsPlugin::PLUGIN_NAME ]),
 					"<br><br>",
 					self::dic()->ui()->factory()->listing()->descriptive($placeholder_types)
 				])
