@@ -9,6 +9,7 @@ use srag\Notifications4Plugins\Exception\Notifications4PluginsException;
  *
  * @package srag\Plugins\Notifications4Plugins\Parser
  *
+ * @author  studer + raimann ag - Team Custom 1 <support-custom1@studer-raimann.ch>
  * @author  Stefan Wanzenried <sw@studer-raimann.ch>
  */
 interface Parser {
@@ -21,7 +22,5 @@ interface Parser {
 	 *
 	 * @throws Notifications4PluginsException
 	 */
-	public function parse(/*string*/
-		$text, array $placeholders = array())/*: string*/
-	;
+	public function parse(string $text, array $placeholders = array()): string;
 }
