@@ -25,20 +25,20 @@ class vcalendarSender implements Sender {
 	use DICTrait;
 	use Notifications4PluginsTrait;
 	const PLUGIN_CLASS_NAME = ilNotifications4PluginsPlugin::class;
-	CONST METHOD_REQUEST = 'REQUEST';
-	CONST METHOD_CANCEL = 'CANCEL';
+	CONST METHOD_REQUEST = "REQUEST";
+	CONST METHOD_CANCEL = "CANCEL";
 	/**
 	 * @var string
 	 */
-	protected $message = '';
+	protected $message = "";
 	/**
 	 * @var string
 	 */
-	protected $subject = '';
+	protected $subject = "";
 	/**
 	 * @var string
 	 */
-	protected $location = '';
+	protected $location = "";
 	/**
 	 * @var ilMail
 	 */
@@ -60,7 +60,7 @@ class vcalendarSender implements Sender {
 	/**
 	 * @var string
 	 */
-	protected $uid = '';
+	protected $uid = "";
 	/**
 	 * @var int
 	 */
@@ -98,12 +98,7 @@ class vcalendarSender implements Sender {
 	 * @param int                  $endTime   Timestamp
 	 * @param int                  $sequence
 	 */
-	public function __construct($user_from = 0, $to = "",  /*string*/
-		$method = self::METHOD_REQUEST, /*string*/
-		$uid = "", /*int*/
-		$startTime = 0, /*int*/
-		$endTime = 0, /*int*/
-		$sequence = 0) {
+	public function __construct($user_from = 0, $to = "", string $method = self::METHOD_REQUEST, string $uid = "", int $startTime = 0, int $endTime = 0, int $sequence = 0) {
 
 		if ($user_from) {
 			$this->setUserFrom($user_from);
@@ -131,7 +126,7 @@ class vcalendarSender implements Sender {
 		//Create Email Headers
 		$mime_boundary = "----Meeting Booking----" . MD5(TIME());
 
-		$this->mailer->sendInternalMail($sent_folder_id, $this->getUserFrom(), '', $this->to, '', '', 'read', 'email', 0, $this->subject, $this->getIcalEvent($mime_boundary), $this->getUserFrom(), 0);
+		$this->mailer->sendInternalMail($sent_folder_id, $this->getUserFrom(), "", $this->to, "", "", "read", "email", 0, $this->subject, $this->getIcalEvent($mime_boundary), $this->getUserFrom(), 0);
 
 		$this->mailer = new ilMail($this->getUserFrom());
 
@@ -143,7 +138,7 @@ class vcalendarSender implements Sender {
 		$headers .= "Content-class: urn:content-classes:calendarmessage\n";
 
 		$result = false;
-		if (!intval(self::dic()->settings()->get('prevent_smtp_globally'))) {
+		if (!intval(self::dic()->settings()->get("prevent_smtp_globally"))) {
 			$result = mail($this->to, $this->subject, $this->getIcalEvent($mime_boundary), $headers);
 		}
 
@@ -239,11 +234,11 @@ class vcalendarSender implements Sender {
 	 * @inheritdoc
 	 */
 	public function reset() {
-		$this->from = '';
-		$this->to = '';
-		$this->subject = '';
-		$this->message = '';
-		$this->uid = '';
+		$this->from = "";
+		$this->to = "";
+		$this->subject = "";
+		$this->message = "";
+		$this->uid = "";
 		$this->method = self::METHOD_REQUEST;
 		$this->sequence = 0;
 		$this->startTime = 0;
@@ -300,7 +295,7 @@ class vcalendarSender implements Sender {
 		$message .= "<html>\n";
 		$message .= "<body>\n";
 		$message .= nl2br($this->message);
-		//$message .= '<p>'.$description.'</p>';
+		//$message .= "<p>".$description."</p>";
 		$message .= "</body>\n";
 		$message .= "</html>\n";
 		$message .= "--$mime_boundary\r\n";
@@ -310,15 +305,15 @@ class vcalendarSender implements Sender {
 			$status = "CANCELLED";
 		}
 
-		$ical = 'BEGIN:VCALENDAR' . "\r\n" . 'PRODID:-//ILIAS' . "\r\n" . 'VERSION:2.0' . "\r\n" . 'METHOD:' . $this->method . "\r\n" . 'BEGIN:VEVENT'
-			. "\r\n" . 'UID: ' . $this->uid . "\r\n" . 'DESCRIPTION:Reminder' . "\r\n" . 'DTSTART;TZID=CET:' . date("Ymd\THis", $this->startTime)
-			. "\r\n" . 'DTEND;TZID=CET:' . date("Ymd\THis", $this->endTime) . "\r\n" . 'DTSTAMP:' . date("Ymd\TGis") . "\r\n" . 'LAST-MODIFIED:'
-			. date("Ymd\TGis") . "\r\n" . 'ORGANIZER;CN="' . $this->from . '":MAILTO:' . $iluser->getEmail() . "\r\n" . 'ATTENDEE;CN="' . $this->to
-			. '";ROLE=REQ-PARTICIPANT;RSVP=TRUE:MAILTO:' . $this->to . "\r\n" . 'SUMMARY:' . $this->subject . "\r\n" . 'LOCATION:' . $this->location
-			. "\r\n" . 'SEQUENCE:' . $this->sequence . "\r\n" . 'PRIORITY:5' . "\r\n" . 'STATUS:' . $status . "\r\n" . 'TRANSP:OPAQUE' . "\r\n"
-			. 'CLASS:PUBLIC' . "\r\n" . 'BEGIN:VALARM' . "\r\n" . 'TRIGGER:-PT15M' . "\r\n" . 'ACTION:DISPLAY' . "\r\n" . 'END:VALARM' . "\r\n"
-			. 'END:VEVENT' . "\r\n" . 'END:VCALENDAR' . "\r\n";
-		$message .= 'Content-Type: text/calendar;name="meeting.ics";method=' . $this->method . "\n";
+		$ical = "BEGIN:VCALENDAR" . "\r\n" . "PRODID:-//ILIAS" . "\r\n" . "VERSION:2.0" . "\r\n" . "METHOD:" . $this->method . "\r\n" . "BEGIN:VEVENT"
+			. "\r\n" . "UID: " . $this->uid . "\r\n" . "DESCRIPTION:Reminder" . "\r\n" . "DTSTART;TZID=CET:" . date("Ymd\THis", $this->startTime)
+			. "\r\n" . "DTEND;TZID=CET:" . date("Ymd\THis", $this->endTime) . "\r\n" . "DTSTAMP:" . date("Ymd\TGis") . "\r\n" . "LAST-MODIFIED:"
+			. date("Ymd\TGis") . "\r\n" . "ORGANIZER;CN="" . $this->from . "":MAILTO:" . $iluser->getEmail() . "\r\n" . "ATTENDEE;CN="" . $this->to
+			. "";ROLE=REQ-PARTICIPANT;RSVP=TRUE:MAILTO:" . $this->to . "\r\n" . "SUMMARY:" . $this->subject . "\r\n" . "LOCATION:" . $this->location
+		. "\r\n" . "SEQUENCE:" . $this->sequence . "\r\n" . "PRIORITY:5" . "\r\n" . "STATUS:" . $status . "\r\n" . "TRANSP:OPAQUE" . "\r\n"
+		. "CLASS:PUBLIC" . "\r\n" . "BEGIN:VALARM" . "\r\n" . "TRIGGER:-PT15M" . "\r\n" . "ACTION:DISPLAY" . "\r\n" . "END:VALARM" . "\r\n"
+		. "END:VEVENT" . "\r\n" . "END:VCALENDAR" . "\r\n";
+		$message .= "Content-Type: text/calendar;name="meeting . ics";method=" . $this->method . "\n";
 		$message .= "Content-Transfer-Encoding: 8bit\n\n";
 		$message .= $ical;
 
