@@ -2,6 +2,7 @@
 
 namespace srag\Plugins\Notifications4Plugins\Utils;
 
+use srag\Plugins\Notifications4Plugins\Notification\Language\Repository as NotificationLanguageRepository;
 use srag\Plugins\Notifications4Plugins\Notification\Repository as NotificationRepository;
 use srag\Plugins\Notifications4Plugins\Parser\Repository as ParserRepository;
 use srag\Plugins\Notifications4Plugins\Sender\Repository as SenderRepository;
@@ -18,15 +19,23 @@ trait Notifications4PluginsTrait {
 	/**
 	 * @return NotificationRepository
 	 */
-	protected static function notification()/*: NotificationRepository*/ {
+	protected static function notification(): NotificationRepository {
 		return NotificationRepository::getInstance();
+	}
+
+
+	/**
+	 * @return NotificationLanguageRepository
+	 */
+	protected static function notificationLanguage(): NotificationLanguageRepository {
+		return NotificationLanguageRepository::getInstance();
 	}
 
 
 	/**
 	 * @return ParserRepository
 	 */
-	protected static function parser()/*: ParserRepository*/ {
+	protected static function parser(): ParserRepository {
 		return ParserRepository::getInstance();
 	}
 
@@ -34,7 +43,7 @@ trait Notifications4PluginsTrait {
 	/**
 	 * @return SenderRepository
 	 */
-	protected static function sender()/*: SenderRepository*/ {
+	protected static function sender(): SenderRepository {
 		return SenderRepository::getInstance();
 	}
 }
